+++ packages/react-native-theming/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @storybook/react-native-theming
 
+## 10.4.4
+
+## 10.4.3
+
+## 10.4.2
+
+## 10.4.2-next.0
+
 ## 10.4.1
 
 ### Patch Changes